requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
djongo==1.3.6
Pillow==8.2.0
pymongo[srv]==3.12.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
text-unidecode==1.3
gunicorn==20.1.0
djangorestframework-simplejwt==5.0.0
chance==0.110